+++ index.html
@@ -227,7 +227,7 @@
     </section>
   </template>
 
-<script src="/js/functions.js"></script>
-<script src="/js/main.js"></script>
+  <script src="/js/main.js" type="module"></script>
+  <script src="/js/training-functions.js"></script>
 </body>
 </html>
